// src/constants.d.ts
export declare const AUTO_DECREASE_AMOUNT = 2;
export declare const AUTO_DECREASE_RATE = 100;
export declare const AUTO_INCREASE_AMOUNT = 10;
export declare const MAXIMUM_THERMOMETER_HEIGHT = 100;
export declare const EXPECTED_MAXIMUM_PERCENTAGE = 100;
export declare const KEYS_TO_HOLD: string[];
export declare const KEY_TO_PRESS = "r";
export declare const LOW_REWARDS: number[];
export declare const MEDIUM_REWARDS: number[];
export declare const HIGH_REWARDS: number[];
export declare const REWARD_OPTIONS: number[][];
export declare const EASY_BOUNDS: number[];
export declare const MEDIUM_BOUNDS: number[];
export declare const HARD_BOUNDS: number[];
// ...
export declare const NUM_DEMO_TRIALS = 1;
export declare const MINIMUM_DEMO_TAPS = 4;
export declare const FAILED_MINIMUM_DEMO_TAPS_DURATION = 1000;
export declare const NUM_TRIALS = 63;
export declare const TRIAL_DURATION = 5000;
export declare const PARAMETER_COMBINATIONS: {
    reward: number[];
    bounds: number[];
}[];
// ...
export declare const FAILED_MINIMUM_DEMO_TAPS_MESSAGE: string;
export declare const HOLD_KEYS_MESSAGE: string;
export declare const KEY_TAPPED_EARLY_MESSAGE: string;
export declare const RELEASE_KEYS_MESSAGE: string;
export declare const RELEASE_KEYS_BACKUP_MESSAGE: string;
export declare const REWARD_TOTAL_MESSAGE: (totalSuccessfulReward: string) => string;
export declare const TUTORIAL_MESSAGE_1: string;
export declare const NO_STIMULI_VIDEO_TUTORIAL_MESSAGE: string;
export declare const STIMULI_VIDEO_TUTORIAL_MESSAGE: string;
export declare const VALIDATION_VIDEO_TUTORIAL_MESSAGE: string;
export declare const INTERACTIVE_KEYBOARD_TUTORIAL_MESSAGE: string;
export declare const COUNTDOWN_DIRECTIONS: string;
export declare const DEMO_TRIAL_MESSAGE: string;
export declare const ACCEPTANCE_TRIAL_MESSAGE: string;
export declare const DOMINANT_HAND_MESSAGE: string;
export declare const TRIAL_FAILED: string;
export declare const TRIAL_SUCCEEDED: string;
export declare const GO_MESSAGE: string;
export declare const LOADING_BAR_MESSAGE: string;
export declare const CONTINUE_BUTTON_MESSAGE: string;