// src/likert.d.ts
import surveyLikert from '@jspsych/plugin-survey-likert';
export declare const likertQuestions1: {
    type: typeof surveyLikert;
    questions: {
        prompt: string;
        labels: string[];
        name: string;
        required: boolean;
    }[];
    randomize_question_order: boolean;
    button_label: string;
}[];
export declare const likertQuestions2: {
    type: typeof surveyLikert;
    questions: {
        prompt: string;
        labels: string[];
        name: string;
        required: boolean;
    }[];
    randomize_question_order: boolean;
    button_label: string;
}[];