// src/stimulus.d.ts
export declare function stimulus(showThermometer: boolean, mercuryHeight: number, lowerBound: number, upperBound: number, error: string): string;
export declare const acceptanceThermometer: (bounds: number[], reward: number) => string;
export declare const videoStimulus: (message: string) => string;
export declare const loadingBar: () => string;
export declare const noStimuliVideo: string;
export declare const stimuliVideo: string;
export declare const validationVideo: string;